progress bar use size_t

+++ client/progress_bar.h
@@ -4,26 +4,26 @@ namespace audit {
 
 class ProgressBar {
  public:
-  ProgressBar(int max) : max_(max) {}
+  ProgressBar(size_t max) : max_(max) {}
 
-  void Progress(int value) {
+  void Progress(size_t value) {
     current_value_ += value;
     percentage_ =
         (static_cast<double>(current_value_) / static_cast<double>(max_)) * 100;
     if (percentage_ > 100) percentage_ = 100;
   }
 
-  int Percentage() const { return percentage_; }
+  size_t Percentage() const { return percentage_; }
 
   bool Done() const { return percentage_ == 100; }
 
  private:
   // When the progress bar reaches this value, it's considered done
-  const int max_;
+  const size_t max_;
 
   // This value measures the current progress relative to max
-  int current_value_{0};
+  size_t current_value_{0};
 
-  int percentage_{0};
+  size_t percentage_{0};
 };
 }
